# Performance boost in _blocks_to_col by np.concatenate