--- a/modin/pandas/utils.py
+++ b/modin/pandas/utils.py
@@ -408,8 +408,7 @@ def _create_blocks_helper(df, npartitions, axis):
 @ray.remote
 def _blocks_to_col(*partition):
     if len(partition):
-        return pandas.concat(partition, axis=0, copy=False)\
-            .reset_index(drop=True)
+        return pandas.DataFrame(np.concatenate(partition, axis=0))
     else:
         return pandas.Series()
 
@@ -417,8 +416,8 @@ def _blocks_to_col(*partition):
 @memoize
 @ray.remote
 def _blocks_to_row(*partition):
-    row_part = pandas.concat(partition, axis=1, copy=False)\
-        .reset_index(drop=True)
+    row_part = pandas.DataFrame(np.concatenate(partition, axis=1))
+
     # Because our block partitions contain different indices (for the
     # columns), this change is needed to ensure correctness.
     row_part.columns = pandas.RangeIndex(0, len(row_part.columns))