--- a/modin/pandas/utils.py
+++ b/modin/pandas/utils.py
@@ -410,7 +410,7 @@ def _blocks_to_col(*partition):
     if len(partition):
         df = pandas.DataFrame(np.concatenate(partition, axis=0))
 
-        if len(np.unique(partition.dtypes.values)) > 1:  # heterogenous types
+        if len(np.unique(partition[0].dtypes.values)) > 1:  # heterogenous types
             df = df.astype(partition[0].dtypes, copy=False)
 
         return df