In `_blocks_to_col`, we can use `np.concatenate` instead of `pd.concat` to speedup our concatenation. Because blocks belongs to the same column group often have the same type or (due to limited col width) few types, the casting will be fast even when it is required.