@simon-mo 2018-07-05
## Quick Benchmark

- *With memoizer turned off*

### Code
```py
import modin.pandas as pd
import numpy as np

types = [np.int]*5 + [np.float]*5
size = int(5e6)

df = pd.DataFrame({
    i: np.random.randn(size).astype(type)
    for i, type in enumerate(types)
})


%%timeit
parts = df._col_partitions
_ = pd.ray.wait(parts, len(parts))
```

### Result

#### Before:

```
The slowest run took 4.25 times longer than the fastest. This could mean that an intermediate result is being cached.
2.14 s ± 1.21 s per loop (mean ± std. dev. of 7 runs, 1 loop each)
```


#### After
```
The slowest run took 4.29 times longer than the fastest. This could mean that an intermediate result is being cached.
1.32 s ± 784 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

#### Note
- The volitality might comes from plasma mmaping files etc. However, with memoizer turned off, numpy concat can be about 2x faster than pandas.